Add Clash.Normalize.Transformations.SeparateArgs
All transformations concerning pulling apart argument types now
live in a separate module. This is part of work to make the
normalization modules quicker to compile.
