+++ clash-lib/clash-lib.cabal
@@ -250,6 +250,7 @@ Library
                       Clash.Normalize.Transformations.Letrec
                       Clash.Normalize.Transformations.MultiPrim
                       Clash.Normalize.Transformations.Reduce
+                      Clash.Normalize.Transformations.SeparateArgs
                       Clash.Normalize.Transformations.Specialize
                       Clash.Normalize.Transformations.XOptimize
                       Clash.Normalize.Types
